+++ manta-pay/src/signer/client/websocket.rs
@@ -22,7 +22,7 @@ use crate::{
     config::{utxo::Address, Config},
     signer::{
         Checkpoint, GetRequest, SignError, SignRequest, SignResponse, SyncError, SyncRequest,
-        SyncResponse,
+        SyncResponse, TransactionDataRequest, TransactionDataResponse,
     },
 };
 use alloc::boxed::Box;
@@ -149,5 +149,13 @@ impl signer::Connection<Config> for Client {
     #[inline]
     fn address(&mut self) -> LocalBoxFutureResult<Address, Self::Error> {
         Box::pin(async move { self.send("address", GetRequest::Get).await })
+    }
+
+    #[inline]
+    fn transaction_data(
+        &mut self,
+        request: TransactionDataRequest,
+    ) -> LocalBoxFutureResult<TransactionDataResponse, Self::Error> {
+        Box::pin(async move { self.send("transaction_data", request).await })
     }
 }