+++ manta-trusted-setup/src/ceremony/registry/csv.rs
@@ -17,8 +17,13 @@
 //! Trusted Setup Ceremony Registry CSV Compatibility
 
 use crate::ceremony::registry::Registry;
+use core::fmt::Debug;
 use openzl_util::serde::de::DeserializeOwned;
-use std::{fs::File, path::Path};
+use std::{
+    fs::{File, OpenOptions},
+    io::{Seek, SeekFrom},
+    path::Path,
+};
 
 /// CSV Record
 pub trait Record<I, V>: DeserializeOwned {
@@ -56,11 +61,58 @@ where
     T: Record<I, V>,
     R: Registry<I, V>,
     P: AsRef<Path>,
+    T::Error: Debug,
 {
     let mut registry = R::new();
-    for record in csv::Reader::from_reader(File::open(path)?).deserialize() {
-        let (identifier, participant) = T::parse(record?).map_err(Error::Parse)?;
-        registry.insert(identifier, participant);
-    }
+    load_append_entries::<_, _, T, _, _>(path, &mut registry)?;
     Ok(registry)
+}
+
+/// Loads new entries into `registry` from `path` using `T` as the record type without overwriting
+/// existing entries. Returns the number of new entries added.
+#[inline]
+pub fn load_append_entries<I, V, T, R, P>(
+    path: P,
+    registry: &mut R,
+) -> Result<usize, Error<T::Error>>
+where
+    T: Record<I, V>,
+    R: Registry<I, V>,
+    P: AsRef<Path>,
+    T::Error: Debug,
+{
+    let length = registry.len();
+
+    for (number, record) in csv::Reader::from_reader(File::open(path)?)
+        .deserialize()
+        .flatten()
+        .enumerate()
+    {
+        match T::parse(record) {
+            Ok((identifier, participant)) => {
+                registry.insert(identifier, participant);
+            }
+            Err(e) => {
+                println!("Line: {} Parsing error {e:?}", number + 2);
+            }
+        };
+    }
+    Ok(registry.len() - length)
+}
+
+/// Build an append-only CSV writer from a file path.
+/// Missing files are created.
+pub fn append_only_csv_writer<E, P>(path: P) -> Result<csv::Writer<File>, E>
+where
+    P: AsRef<Path>,
+    E: From<std::io::Error>,
+{
+    let mut file = OpenOptions::new().write(true).create(true).open(&path)?;
+    let mut file_is_empty = false;
+    if file.seek(SeekFrom::End(0))? == 0 {
+        file_is_empty = true;
+    }
+    Ok(csv::WriterBuilder::new()
+        .has_headers(file_is_empty)
+        .from_writer(file))
 }